CleanUp for OnBallHitPlayerEvent

+++ Scripts/Events/BallHitPlayerEvent.cs
@@ -2,4 +2,4 @@ using pong_1.Scripts.EventBus;
 
 namespace Pong_1.Scripts.Events;
 
-public record BallHitPlayerEvent(float collisionY, int playerHeight, float playerPosition) : IEvent;
+public record BallHitPlayerEvent(float collisionY, float playerHeight, float playerPosition) : IEvent;
